LocalRepository: Re-implement methods in GithubPackages using a LocalRepository object
