+++ Snowflake.Packaging/Program.cs
@@ -109,20 +109,10 @@ namespace Snowflake.Packaging
                     }
                     else
                     {
-                        Task.Run(async () =>
+                        foreach (var dep in manager.PackageRepository.ResolveDependencies("testplugin"))
                         {
-                            try {
-                                var deps = await GithubPackages.ResolveDependencies(options.PackageFiles[0]);
-                                foreach (var dep in deps)
-                                {
-                                    Console.WriteLine(dep.Name);
-                                }
-                            }
-                            catch (Octokit.ApiException)
-                            {
-                                Console.WriteLine("Rate limit exceeded, log in with GitHub credentials using snowball auth");
-                            }
-                        }).Wait();
+                            Console.WriteLine(dep.Name);
+                        };
                     }
                     //todo wrap in try/catch
