+++ Snowflake.Packaging/Snowflake.Packaging.csproj
@@ -70,7 +70,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Commands.cs" />
-    <Compile Include="Installing\GithubPackages.cs" />
     <Compile Include="Installing\InstallManager.cs" />
     <Compile Include="Installing\LocalRepository.cs" />
     <Compile Include="Properties\Settings.Designer.cs">
